+++ src/main/java/org/truffleruby/language/RubyContextSourceNode.java
@@ -23,7 +23,7 @@ public abstract class RubyContextSourceNode extends RubyNode {
 
     @Override
     public Object isDefined(VirtualFrame frame, RubyLanguage language, RubyContext context) {
-        return RubyNode.defaultIsDefined(getLanguage(), context, this);
+        return RubyNode.defaultIsDefined(this);
     }
 
     @Override