+++ src/main/java/org/truffleruby/language/constants/ReadConstantWithLexicalScopeNode.java
@@ -13,6 +13,7 @@ import org.truffleruby.RubyContext;
 import org.truffleruby.RubyLanguage;
 import org.truffleruby.core.module.ModuleOperations;
 import org.truffleruby.core.module.RubyModule;
+import org.truffleruby.core.string.FrozenStrings;
 import org.truffleruby.language.LexicalScope;
 import org.truffleruby.language.RubyConstant;
 import org.truffleruby.language.RubyContextSourceNode;
@@ -55,7 +56,7 @@ public class ReadConstantWithLexicalScopeNode extends RubyContextSourceNode {
         }
 
         if (ModuleOperations.isConstantDefined(constant)) {
-            return coreStrings().CONSTANT.createInstance(getContext());
+            return FrozenStrings.CONSTANT;
         } else {
             return nil;
         }