+++ src/main/java/org/truffleruby/language/exceptions/RescueClassesNode.java
@@ -41,7 +41,7 @@ public class RescueClassesNode extends RescueNode {
 
     @Override
     public Object isDefined(VirtualFrame frame, RubyLanguage language, RubyContext context) {
-        return RubyNode.defaultIsDefined(getLanguage(), context, this);
+        return RubyNode.defaultIsDefined(this);
     }
 
 }