+++ src/main/java/org/truffleruby/language/exceptions/RescueStandardErrorNode.java
@@ -29,7 +29,7 @@ public class RescueStandardErrorNode extends RescueNode {
 
     @Override
     public Object isDefined(VirtualFrame frame, RubyLanguage language, RubyContext context) {
-        return RubyNode.defaultIsDefined(getLanguage(), context, this);
+        return RubyNode.defaultIsDefined(this);
     }
 
 }