+++ src/main/java/org/truffleruby/language/supercall/SuperCallNode.java
@@ -11,6 +11,7 @@ package org.truffleruby.language.supercall;
 
 import org.truffleruby.RubyContext;
 import org.truffleruby.RubyLanguage;
+import org.truffleruby.core.string.FrozenStrings;
 import org.truffleruby.language.RubyContextSourceNode;
 import org.truffleruby.language.RubyNode;
 import org.truffleruby.language.arguments.RubyArguments;
@@ -58,7 +59,7 @@ public class SuperCallNode extends RubyContextSourceNode {
         if (superMethod == null) {
             return nil;
         } else {
-            return coreStrings().SUPER.createInstance(context);
+            return FrozenStrings.SUPER;
         }
     }